A proposal to deprecate APIs we never intended
There's a lot of cruft in `slang.h` and the `slangc` command-line options that was never really meant to be used, and a lot of other cruft that we knew was a liability from the moment it landed.
This is a proposal to mark as much stuff deprecated as currently seems possible, with the hope that over time we can clean up the implementation by not needing to support too many legacy features and in-retrospect-poor design choices.